Core files in place
I am in the process of porting over earlier attempts at the engine. This is like the initial commit of some the Core files.

There is a TestGame project that just prints out some logs. I aim to have this as a demo project when I get more things in.