Removed ProjectGenerator from the Solution Build!
I had an issue where I was Building the Project Generator (which in turn rebuilds the solution) as a part of the solution build! Thanks to stack overflow, I was able to add a "buildenabled" command as part of the build system to remove it from the builds in all configurations! Neat!