Finished Triangles for now
- I finished the Triangles chapter in my book, and added the Area, Signed Area, and another implementation (though not used right now) for calculating a Barycentric coordinate.
- Added `ToString()` as well for the triangle classes.
- Added the maybe_unused flag to remove the warning in the WorldManager for now. This whole class needs to be reevaluated later.

Something I found out: even if you have the conversions written in code between types, the compiler seems to be only able to do 1 of the conversions.   I had written a conversion from a 2D Vector2 to a Vector3, but because the type of the Components also needed to be defined, it didn't want to do both conversions and reported an error. I made explicit functions for handling 2D and 3D vertices, and that fixed the issue. I feel like I haven't run into that before, so I may have been doing something wrong... Something to double check when you can.