ECS Rework, enforcing Component ownership types.
I ran into an issue when making the FreeCamMovementComponent. While it did not exist in 3D space, meaning it didn't need to inherit from WorldComponent, it *was* expecting that the Owning type does exist in 3D space. The options were to perform a cast on every time we update our position, or, which is what I went with, enforcing an OwnerType for components.

- Entities and EntityLayers got reskinned to SceneNodes and SceneLayers. I could write a whole diatribe on the naming fiasco, but I will leave that out.

- Components are now defined by an owning Type. You cannot add a component to a SceneNode of a different type. While this kind of sucks in the case that I write a "HealthComponent" that could theoretically attach to either a 3D Node type or a 2D Node type, it helps with the overall API. That was the tradeoff that I went back and forth with.

- Abstract Layer is removed for now. I am focusing on the 3D world for now and I am going to concentrate on that as best I can.

- Added a FreeCamMovementComponent

- Updated the test scene to load the new components.