Base ECS Structure is finally done.
I simplified things. I was infatuated with the idea of a RootComponent for Entities in the way that Unreal handles Actors vs Components. I scrapped that idea. The goal was to restrict component and Entity attachment based on a domain. By having the Components take in an owner type, that defines their domain. All of the parenting and transform information is now on the derived Entity, rather than in a Component.

An Entity is something that exists in a domain. A Component is something that can attach to a specific Entity Type that can give it functionality.

- Actor renamed to Entity3D.
- World Component removed; all transform information is directly on the Entity3D class.
- Component registration moved into TEntity base class.