Basic Nouns, the Quad Tree
This commit contains "stubbed-in" classes (functions have no implementation). The Physics classes are not used in the codebase, so everything builds just fine.

The reason for this commit is for school, so that I can show the work that I have been doing on this branch.

I have been spending most of my time implementing the QuadTree. This will be used in the broadphase of the physics sim, to reduce the amount of work in the Narrow Phase.

Terms:
- "Body" : This contains information about the physics object, velocity, angular velocity, etc as well as determining an overall bounding box (AABB). An 3D Entity will have a BodyID which will be in the Physics System.
- "Shape": This is the different bounding shapes that a Body can have. Use in Narrow Phase intersections.
- "Constraint": Defines how Bodies are constrained to one another - that is specifically a TwoBodyConstraint. This will be subclassed for behavior such as Hinges, having an object be "welded" to another, etc.
- "CollisionLayer": Numerical value that Bodies can be in, which can be used to filter out collisions between different Bodies.
- "BroadPhaseLayer" : Numerical value that determines which broadphase tree the Body should be in. Bodies in separate trees will never collide with one another. Collision Layers can be mapped to a broadphase layer.
- "PhysicsScene": Class that runs the Physics Simulation. Bodies are registered to this class to be part of the sim.