Base Shape, Body, other Physics Classes
Huge commit with the base Shape & Empty Shape classes implemented, Body and Body Manager are implemented, tons of property classes involving both, and the beginnings of the Narrow Phase with the Collision Solver class.

There is an absolute boat load of code in this update. I wanted to commit earlier, but the types are very dependent which made it difficult.