Successfully added a Body with a Box!
The game still doesn't show the effects of the physics system running, but I have added a simple test for adding/removing a Body with a Box Shape attached to the system. I have verified that the box is successfully added to the broad phase.

Body Interface Class:
- This is the public-facing interface for creating/destroying/updating Body's and their shapes. The World will be using that when registering/removing Entity's Shape Components and the like. The interface entirely works through BodyIDs, which acts as a handle to the underlying Body object.
- The interface takes in a BodyLockInterface, to have an interface that locks for every access vs a non-locking version.

I have also begun work on Constraints, which defines contacts between objects, or how objects are attached to one another. This is crucial for the Narrow Phase.