The Narrow Phase Query is Complete!
Turns out I had to do a bit more than I thought for the Internal Edge Removing Collector, because of the STL Allocator types. One is a wrapper to use my allocation functions, and the other maintains a local buffer before deferring to heap allocations when that buffer is full.

With the Narrow Phase Query sorted, I will be working on the Physics Scene again! It's been far too long 😅.