Added NRI for Graphics API Abstraction.
I have decided to use the NRI library to handle the Vulkan Abstraction that I was starting (I was starting to basically remake the same thing). I will then focus on making the C++ wrapper for the NRI.
- I have removed VkBootstrap.

Because NRI uses cmake to build the project, I had to recreate parts of the build process to be suitable with my build system that uses premake.
Some remaining issues:
- D3D11/12 are not supported right now. There is a bunch of work to do to get those libraries ready, and I just want to support Vulkan anyway.
- The ShaderMake files are in, but not completely setup correctly.
- I took the build output of NRI and just put the relevant files that I needed straight into the NRI folder (_deps & _Shaders folders). I need to automate the process when I can.

Good news is that I have fixed up the virtual paths (vs "filters") throughout the project as a side product of the work today. Which is nice.