Vulkan Instance Created, Physical Device Selected!
- Updated RendererDesc with more options about queue availability.
- Instance created and shutdown properly
- Correctly selected a physical device.

The Physical Device selection isn't entirely done - I need to set the queue family indices and ensure that extensions are supported, and I need to fill out all of the Physical Device info.

Surface creation will be handled by the Swapchain object, so that will wait for now.

Next are the Queue families, and then the Logical Device.