Device Object Refactor
A Device Object (formerly Device Asset) contains references to one or more resources on the GPU.

They are designed as RAII classes ("Resource Acquisition is Initialization"), meaning that their constructor performs initialization, and their destructor will release the GPU resource. So when the Device Object leaves scope, its resource will be freed!

Device Objects are *not* copyable. They are only movable.

Device objects can be constructed with and assigned to nullptr. Constructing with nullptr is effectively the default constructor. Assigning an initialized Device Object to nullptr will free the resource.

The Render Device no longer contains the templated Create/Free Resource API. Device Objects take in a reference to the Render Device object on construction which will create the necessary resource.  I am no longer dynamically allocating a new object for every resource (in addition to the allocation for the graphics resource).  The previous API also made it possible for a Device Queue to be created outside of the Render Device; same story for Command Buffers which should only be allocated by the Command Pool. Now those constructors are protected.

The Device Object's destructor adds the resource to the free queue on the Renderer to ensure that they are freed once they are no longer in use.