Index Buffers, Staging Buffers
The sample app changed from 'SimpleTriangle' to 'Rectangle', which demonstrates how to make a draw call that utilizes an index buffer and how to upload data to a device only resource.

I put in an interface to create and submit a temporary command buffer that I needed in order to perform the data upload in the Application's Init function.