DeviceBufferRange, Updated Vertex/Index Buffers
I realized that vertex and index buffers are just views into a device buffer resource. So, I created the `DeviceBufferRange` class, which is analogous to a std::span. The `VertexBufferRange` and `IndexBufferRange` are subclasses that contain more specific information about the range and can be used for draw calls.

Having these terms makes it more clear that you can use a single `DeviceBuffer` for multiple ranges of data.