Started Mip Map Generation
I went through and implemented a image blitting method from the Vulkan tutorial, however, image blitting needs to be submitted to a graphics queue (I am generating mip maps on the Asset Thread, which will be using a transfer queue) and blitting is not supported by all image formats.

The solution is to generate the mip maps beforehand and store them in Texture Asset file itself, and then upload data to each level accordingly.

I removed access to it, but I left the GenerateMipMaps function in there for reference in the future.

I added some comments to the image description object, and documented what I need to do later.