Instance Data in Push Constants
I move the Instance data (model matrix, normal matrix) into the Push Constants and created a storage buffer for the Material Params. In the future, I would have the Instance data in another storage buffer and a compute shader will build draw commands based on the objects within the frustum.