Render Targets, Data-Driven Pipelines
Removed the GBuffer class in favor of "Render Targets". The Render Target API resembles the GBuffer, only it pertains to a single image rather than a set of color images and optional depth/stencil image, which was already represented by the Render Targets Desc that is sent to the Command Buffer - I had two classes that were similar in structure.

Now the Render Targets are resized individually. They also maintain their format, clear values, image planes and debug names. I believe it makes the API more clear.

In the PBR project, I now load the render targets and graphics pipeline description from data. For the color/depth attachments, they have a Render Target field that contains the name of the target to use, which allows me to set the format and sample count based on the Render Target data.