Camera Settings data driven.
Added Camera Settings to the file containing the render target and pipeline information, then renamed the file to PBRAppSettings. Moved it to the Config directory.

Removed the Mesh and Material classes from Nessie, as they are specific to the graphics pipeline used. After the semester, I might try to make this PBR project a library that can be used for application projects. An example pipeline that can be used. It will have the Mesh and Material Components specific to the pipeline.