Implemented Cascading Shadow Maps!
I now do a depth pass, rendering the scene objects from the perspective of the directional light in different levels of detail. These maps are sampled in the PBR shader to shadow areas of the scene. Based on the proximity of the vertex position to the camera, a different level of detail shadow map is used.