ECS System in!
The new ECS System is pretty much complete!
- Largely a wrapper for the entt library.
- Components are just data.
- Entities are just an ID.
- EntityRegistry contains all Entities and their Components. Manages creating and destroying both.
- A World is contains EntityRegistry and an array of Component Systems to operate on those Entities.
- A Component System processes a group of entities based on which components they have.
- Component Registry manages serialization functions when loading a WorldAsset. Component Systems register the Component types that they need to use.
- A World Asset maintains its own EntityRegistry. It can be merged into the runtime World.

The only thing left to do is have the World Asset load all Assets that the World needs as well. I plan on making a container of AssetID-filepath pairs that can be loaded on a discrete step.

I also removed the Scene and World code that this refactor replaces.