AssetPacks are in!
An Asset Pack is an array of AssetMetadata, which contains an AssetID, a TypeID, and a file path. They can be loaded directly with the AssetManager, both synchronously and asynchronously. The WorldAsset has an AssetPack, which is a set of Assets that need to be loaded for the World to be able to run. The Assets aren't loaded when the WorldAsset is loaded - they have to be loaded in a separate step. This is in anticipation of transitioning between worlds, or being able to load sections of a World.

- Asset Types can now be registered to the AssetManager, to be able to be loaded through AssetMetadata rather than knowing the asset's type directly.
- The AssetMetadata struct is now used throughout the AssetManager, which allowed me to cut back on templated code!
- The DataUploader now passes DeviceQueue ownership to the Render DeviceQueue if uploading on the AssetThread.