Fixed Async Asset Loading for Textures!
I was having an issue where the image layouts weren't be properly being transitioned when loaded on the Asset Thread. Since I am using a dedicated queue, I had to set up a 'release' barrier for the Transfer Queue, and an 'acquire' barrier for the render queue. BUT, I didn't realize that I had to match the layout transition for both barriers - I thought that I just had to transition the layout on the release barrier. That was a ton of debugging, but I got it.