GetNumEntities
You can now query the number of entities in the registry. I am using this to only update the world if there are entities present. This was to make sure that the day night system doesn't update and start the sun at a different angle depending on the load time.