Removed Platform class
Applications are going to run Worlds. The Worlds are the design point for developers. The Application class's behavior is going to be defined by the Engine, in either a editor or runtime version.

Previously, the 'Platform' ran 'Applications' which ran 'Worlds'. So I got rid of one of those levels for clarity and simplicity.