Removed imgui.ini, Basic YAML Conversions
I realized I added the imgui.ini file, which is just user file that shouldn't be saved. It is now stored in "Intermediate/ProjectFiles/" folder, which is not uploaded to git.

I started work on a more formal serialization framework - I've added some YAML conversion code. A lot more needs to be added.