Saving the World Asset!
I am now successfully saving the World asset after making changes made in the Editor! I have the basic ability to reparent the Entities in the hierarchy. I still need to be able to reorder and un-parent the entities completely, that is up next.

For now, worlds are responsible for dealing with parenting calls. This is done through a virtual function. In the future, I may just have some Transform System base class that is required on the World object.