Fixed Forward Declared Types
I previously had the issue where types that were forward declared couldn't be used with my StrongPtr. This is obviously horrific. The fix was to change the 'IsRefTargetType' concept to a type trait, since it was used as part of the destructor of the StrongPtr. Concepts must be evaluated at compile time, and since the type was invalid, it caused compilation errors. The type_trait, on the other hand, can fail silently when the type is unknown (SFINAE).