PBR Project updated! Simulation Controls
The PBR project now has the editor interface set up and I can now play and pause the PBR World in the editor!

There are a few visual errors to fix, having to do with resizing and the orthographic camera, but overall this is a huge win! There was a ton of bug fixing and work done to get it to this state!