Viewport Resizing Bugs Fixed!
- Fixed a bug where the window being resized to the right will all of a sudden cause the viewport image to be offset to the left.
- Fixed a bug where the viewport was rendering incorrectly when fullscreen.

Both of these had to do with my resizing logic with the viewport. The viewport is now resizing dynamically. I also had to fix an issue with the render order; I was previously trying to use an image after it was resized later in the hierarchy, causing it to be invalid and start flickering.

I also added a similar interface to Unity for changing the aspect ratio of the viewport.

Viewport controls no longer render when simulating the game. It was a bit confusing considering a number of the controls effected the Editor Camera, not the in game camera.