Sci-fi Corridor In!
I have a simple sci-fi corridor (with no doors or ceilings...) and a new space skybox to act as my demo scene. I am going to add a bit more most likely, but I want to check in with the teacher to see if this is adequate for now.

In the meantime, I am going to focus on other bugs and polish.