Updated Naming in Hierarchy Window
When a new entity is created, I added a counter value to the end of the name, just to give the name more a unique ID. Similar to Unity and Unreal.

I now also allow spaces in the name, but discard a leading space. Just feels more clean to me.