Custom Title Bar Fixed!
I now have my custom title bar working, with windows OS animations, window snapping, etc. It was a long journey, but I finally found a solution that works! Stoked! I will have to make a custom fork of GLFW when I can with these changes.

Window changes (position, minimize, maximize, resize, etc) are deferred to the next frame's update to be applied. This was causing a crash when moving the window between monitors.