Fixed Material Asset Invalid Crash
In the editor, if you set a material to the "None" case, it would cause a crash because the invalid material AssetID would not be added to the materials needed for the mesh in the scene renderer. I fixed this by giving the Property fields for Assets the option to set a default case, which allowed me to default to the imported materials as a default case.